fix: pdf toc when using with-clicks option
