+++ packages/types/src/index.ts
@@ -1,3 +1,4 @@
 export * from './types'
 export * from './setups'
 export * from './config'
+export * from './toc'
